requirements.txt
crewai==0.119.0
crewai_tools>=0.44.0
python-dotenv>=1.0.0
litellm>=1.34.10
python-dotenv
PyMuPDF==1.22.3
pandas
openai
google-generativeai
pyyaml
openpyxl
markdown
pdfkit
python-docx
litellm
